WIP - moved activity base and operation result to core
